Milestone: CPU65XX Instruction Set Complete
Reached a major milestone! All the CPU65XX are present and appear to be
functioning correctly based on the CPU test ROM. There is still a minor
discrepancy at the end of the CPU test, which may be due to how memory
is initialized. Will look more into it.

* Fixed several issues with Combined ALU opcodes.
* Fixed where the CPU test program halts.

Next milestone? A functional PPU.